requirements.txt
Flask==2.0.1 
Pillow==9.0.0 
torch==1.10.0 
torchvision==0.11.1 
matplotlib==3.5.1 
numpy==1.21.4